chore: refactor transfers to use timeline data
