+++ bc_obps/registration/api/v2/__init__.py
@@ -6,7 +6,6 @@ from . import (
     facilities,
     naics_codes,
     operations,
-    operations_timeline,
     operators,
     regulated_products,
     reporting_activities,
