+++ bc_obps/registration/schema/v2/operation.py
@@ -6,7 +6,7 @@ from registration.schema.v1.multiple_operator import MultipleOperatorOut
 from registration.models.contact import Contact
 from registration.schema.v1.operator import OperatorForOperationOut
 from registration.schema.v2.multiple_operator import MultipleOperatorIn
-from ninja import Field, FilterSchema, ModelSchema, Schema
+from ninja import Field, ModelSchema, Schema
 from registration.models import MultipleOperator, Operation
 from registration.models.opted_in_operation_detail import OptedInOperationDetail
 from pydantic import field_validator
@@ -215,38 +215,6 @@ class OperationUpdateOut(ModelSchema):
     class Meta:
         model = Operation
         fields = ['id', 'name']
-
-
-class OperationFilterSchema(FilterSchema):
-    # NOTE: we could simply use the `q` parameter to filter by related fields but,
-    # due to this issue: https://github.com/vitalik/django-ninja/issues/1037 mypy is unhappy so I'm using the `json_schema_extra` parameter
-    # If we want to achieve more by using the `q` parameter, we should use it and ignore the mypy error
-    bcghg_id: Optional[str] = Field(None, json_schema_extra={'q': 'bcghg_id__id__icontains'})
-    name: Optional[str] = Field(None, json_schema_extra={'q': 'name__icontains'})
-    type: Optional[str] = Field(None, json_schema_extra={'q': 'type__icontains'})
-    status: Optional[str] = Field(None, json_schema_extra={'q': 'status__icontains'})
-    bc_obps_regulated_operation: Optional[str] = Field(
-        None, json_schema_extra={'q': 'bc_obps_regulated_operation__id__icontains'}
-    )
-    operator_id: Optional[UUID] = Field(None, json_schema_extra={'q': 'operator__id__exact'})
-    operator__legal_name: Optional[str] = Field(None, json_schema_extra={'q': 'operator__legal_name__icontains'})
-
-
-class OperationListOut(ModelSchema):
-    operator__legal_name: str = Field(..., alias="operator.legal_name")
-    sfo_facility_id: Optional[UUID] = Field(None, alias="sfo_facility_id")  # this is an annotated field in the query
-    sfo_facility_name: Optional[str] = Field(None, alias="sfo_facility_name")  # this is an annotated field in the query
-    bcghg_id: Optional[str] = Field(None, alias="bcghg_id.id")
-
-    class Meta:
-        model = Operation
-        fields = [
-            'id',
-            'name',
-            'type',
-            'status',
-            'bc_obps_regulated_operation',
-        ]
 
 
 class OperationCurrentOut(ModelSchema):
