+++ bciers/apps/administration/app/components/operations/OperationDataGrid.tsx
@@ -8,7 +8,7 @@ import OperationFacilitiesActionCell from "apps/administration/app/components/op
 import operationColumns from "../datagrid/models/operations/operationColumns";
 import operationGroupColumns from "../datagrid/models/operations/operationGroupColumns";
 import { OperationRow } from "./types";
-import fetchOperationsTimelinePageData from "./fetchOperationsTimelinePageData";
+import { fetchOperationsPageData } from "@bciers/actions/api";
 
 const OperationDataGrid = ({
   initialData,
@@ -45,7 +45,7 @@ const OperationDataGrid = ({
     <DataGrid
       columns={columns}
       columnGroupModel={columnGroup}
-      fetchPageData={fetchOperationsTimelinePageData}
+      fetchPageData={fetchOperationsPageData}
       paginationMode="server"
       initialData={initialData}
     />
