--- bciers/apps/administration/app/components/operations/fetchOperationsTimelinePageData.tsx
@@ -1,24 +1,0 @@
-import buildQueryParams from "@bciers/utils/src/buildQueryParams";
-import { OperationsSearchParams } from "./types";
-import { actionHandler } from "@bciers/actions";
-
-// 🛠️ Function to fetch operations
-export default async function fetchOperationsTimelinePageData(
-  searchParams: OperationsSearchParams,
-) {
-  try {
-    const queryParams = buildQueryParams(searchParams);
-    // fetch data from server
-    const pageData = await actionHandler(
-      `registration/operations-timeline${queryParams}`,
-      "GET",
-      "",
-    );
-    return {
-      rows: pageData.items,
-      row_count: pageData.count,
-    };
-  } catch (error) {
-    throw error;
-  }
-}
