+++ bciers/apps/administration/tests/components/operations/mocks.ts
@@ -5,7 +5,6 @@ const getReportingActivities = vi.fn();
 const getRegulatedProducts = vi.fn();
 const getRegistrationPurposes = vi.fn();
 const getBusinessStructures = vi.fn();
-const mockFetchOperationsTimelinePageData = vi.fn();
 const fetchOperationsPageData = vi.fn();
 
 vi.mock("libs/actions/src/api/getOperation", () => ({
@@ -39,12 +38,6 @@ vi.mock("libs/actions/src/api/getBusinessStructures", () => ({
 vi.mock("libs/actions/src/api/fetchOperationsPageData", () => ({
   default: fetchOperationsPageData,
 }));
-vi.mock(
-  "apps/administration/app/components/operations/fetchOperationsTimelinePageData",
-  () => ({
-    default: mockFetchOperationsTimelinePageData,
-  }),
-);
 
 vi.mock(
   "apps/administration/app/components/operations/fetchOperationsPageData",
@@ -62,5 +55,4 @@ export {
   getRegistrationPurposes,
   getBusinessStructures,
   fetchOperationsPageData,
-  mockFetchOperationsTimelinePageData,
 };
