+++ bciers/apps/registration/app/components/transfers/TransferForm.tsx
@@ -17,8 +17,9 @@ import { FacilityRow } from "@/administration/app/components/facilities/types";
 import TaskList from "@bciers/components/form/components/TaskList";
 import { actionHandler } from "@bciers/actions";
 import TransferSuccess from "@/registration/app/components/transfers/TransferSuccess";
-import fetchOperationsPageData from "@bciers/actions/api/fetchOperationsPageData";
-import { OperationTransferData } from "@/administration/app/components/operations/types";
+import { OperationRow } from "@/administration/app/components/operations/types";
+import { fetchOperationsPageData } from "@bciers/actions/api";
+
 interface TransferFormProps {
   formData: TransferFormData;
   operators: OperatorRow[];
@@ -95,7 +96,7 @@ export default function TransferForm({
   const fetchOperatorOperations = async (operatorId?: string) => {
     if (!operatorId) return [];
     const response: {
-      rows: OperationTransferData[];
+      rows: OperationRow[];
       row_count: number;
     } = await fetchOperationsPageData({
       paginate_results: false,
@@ -172,8 +173,8 @@ export default function TransferForm({
     if (!error) {
       // Filter out the current from_operation from toOperatorOperations(we can't transfer facilities to the same operation)
       const filteredToOperatorOperations = toOperatorOperations.filter(
-        (operation: OperationTransferData) =>
-          operation.id !== formState?.from_operation,
+        (operation: OperationRow) =>
+          operation.operation__id !== formState?.from_operation,
       );
 
       setSchema(
