+++ bciers/apps/registration1/e2e/poms/operations.ts
@@ -375,11 +375,11 @@ export class OperationsPOM {
               OperationStatus.APPROVED,
               OperationStatus.DECLINED,
               OperationStatus.CHANGES_REQUESTED,
-              OperationStatus.REGISTERED,
-              // Below values are not possible in Registration1
-              OperationStatus.CLOSED,
-              OperationStatus.TEMPORARILY_SHUTDOWN,
-              OperationStatus.TRANSFERRED,
+              // OperationStatus.REGISTERED,
+              // // Below values are not possible in Registration1
+              // OperationStatus.CLOSED,
+              // OperationStatus.TEMPORARILY_SHUTDOWN,
+              // OperationStatus.TRANSFERRED,
             ];
             break;
         }
